Add "$ref" for subschemas to meta-schemas.
This ensures that subschemas are either "normal" JSON Schema
objects according to the declared vocabulary, or references.

The separate definitions of jsonReference and notJsonReference
ensure that objects with a "$ref" property that is not a URI
reference fail validation.  If we just negated the jsonReference
defintion, then {"$ref": 4} would validate as a non-reference
schema which is incorrect.