Show that we are an active project
The front page doesn't really make it clear that the project
is as active as it is.  Let's put up some actual text on that,
and link to where people can see the status and timeline of
the various drafts.  And a link to the current active milestone,
which will need to be kept up to date but twice a year isn't
exactly burdensome.  And if it's out of date people will see
that the milestone is complete and figure it out anyway.