Rework emphases throughout the page.
Emphasize the initial link to the migration notes page a bit more,
but leave the section further down as people still often miss
links in text.

Rework the meta-schema instructions as all-bold was too heavy
and obscured the important parts of the sentence.

Change "Other versions" to "Older drafts" to be more clear about
what the link points to.  There are not other versions in the
sense of other currently valid versions.