Add ready-for-review works in progress
These will be the "draft-07" documents.  They will be modified
in a subsequent commit to clearly indicate that they are for
review only, and not for use.

The "WIP-" prefix hopefully also clues people in that there
is something unusual going on here.

Doing everything possible (in this and the next commit)
to make it clear that these are not published files, but rather
are pre-built review copies for those who are unlikely to build
their own from the github repo.