Also copy draft-07 docs to that directory
So that we can have links to the latest draft-07 docs
instead of to URIs that will at some point switch to draft-08.

Note that if we have a bugfix draft-handrews-*-01 release that
does not change meta-schemas, that release will be copied over
here as well.  This is not an archive of all published drafts.