Move HTML of draft-01 into dir like others
While there's not much we can do with the .txt or .xml files in
the top level directory right now, we can make the .html use
a redirect so we can put it in its draft-NN directory as we
do with the more recent drafts.

For the .txt and .xml files, while you can get Jekyll to process
them, the redirect plugin is implented by emitting an HTML file
with a canonical link header, as there is no access to the web
server itself to respond with an HTTP redirect code.