Standardize migration notes, improve layout
Standardized on "release notes" rather than FAQ.

Left redirects for the old URIs.

Add a TOC to the specification page as it was getting confusing,
leading to migration notes in multiple places in an attempt to make
them more visible.

Removed out of date "will soon be updated" text, and added links
to the draft-06 notes as well as they have lots of useful
information not replicated in the draft-07 notes.