Link to issues/branches for draft progress
Since the "modern" data set focuses on draft-06+, make a note
of where progress towards that support can be tracked for projects
that are actively working towards it.  We do not want to penalize
in-progress work, we only want to separate out projects that
are either explicitly refusing to move past draft-04, abandoned,
or just hoping someone new shows up and adds support without any
active plan to do so.