Merge CLI/web validators into main section
Having these separate took up a lot of space and also made things
appear in inconsistent ways.

Just treat them as "languages" (which we do in other sections
anyway), and remove the sorting by language name so we can
keep them at the end.

Also compress the draft support wording/formatting as it was
taking up a lot of space.

All notes should be unformatted (not italicized)

Don't show empty () if there's no license given.

One web tool was not even a validator so move it to the right section.

A Go tool from another section is also a validator, so add it.