Reorganize non-validator implementations
Group by use case.  Some projects were mis-categorized, and the
"Other" section was kind of a mess.

Also, define what non-validators/non-hyperclients need to do to
be considered compatible with draft-06 or later and remain on
this page.

Use #### headings as otherwise the difference between the
undlerline headings and the ### headings is not sufficiently
visually distinctive.