Add a status update and more about getting to RFC
Particularly since two of the four drafts are expired right now,
and at the current rate we probably won't quite manage to publish
another round before the other two expire, it seems like a good
idea to make it very clear that the project is active.

That should soon be apparent from PRs, but it's been a bit
slow over the summer.