Redirect the standard vocabulary URIs
We've already had one person confused that these URIs 404, even
though the spec says to expect that.  Let's generate HTML stubs
to redirect to the appropriate spec section for each vocabulary
(or the whole spec for Hyper-Schema).

No media type restrictions are given at this time for vocabulary
documents, so having HTML here is fine.  Jekyll (at least the
local version, production sometimes differs) will do the redirect
on URIs without the .html suffix, which is what we need.  It will
also do it with .html, but I think that's fine.