Redirect unversioned meta-schemas
This replaces these with HTML stubs that will use a meta refresh
tag and canonical link to perform a redirect.

If there's code out there assuming that these exist and are JSON,
then that code will be confused by receiving HTML.  Although we do
say that you can't assume these things are actually hosted.
Not sure this is what anyone had in mind, though.