Schema for the impls data file, structured page
This does two things in preparation for showing more
information on this page in a way that is easy to see
when quickly skimming the list.

It adds a schema (in yaml because it will get pulled
in to the jekyll/liquid template in the future) which
documents the current fields.

It also makes the rendered form of the list more
structured by giving each implementation a field list.
I anticipate more fields as important information that
is currently inconsistenlty reported in the notes gets
pulled out, starting with the configuration compliance
information in the next commit.

This commit also fixes where the language-level notes
were not being rendered correctly, and updates the draft
list for ajv-cli because I happened to notice it was
out of date.

For hyper-schema, the updated layout is simpler because
there is only one modern implementation anyway.
The other was moved to obsolete following the similar
recent moves of validators that have not followed
through on draft-06 or newer support.

The obsolete data files are supported by changing a
handful of license values from arrays to strings,
which display identically.  Allowing arrays didn't
seem worth the trouble since the modern file doesn't
use them.  I did not change the layout of the obsolete
page because really we'd rather people not look at it
at all.  Also, one obsolete validator did not specify
which draft it supported, but from the last modified
date in the repo it must have been draft-03.  At this
point, I doubt it matters anyway.