Add a compliance section to the impl data
This adds a structured section on compliance, specifically how
to configure for it.  This is *not* intended to be the same sort
of compliance information that the test suite provides.  Rather,
it is about things that implementations document that don't
necessarily show up in test suite results.

This format includes brief instructions and optionally
(ideally) a documentation link.  I think this langauge is
appropriately neutral, and particularly when there is a doc link
it feels more like calling attention to something that is already
advertised rather than being judgemental.

The AJV and Opis cases are straightfoward.

The Common Lisp implementation's case of needing to disable
automatic downloading is worth a bit of thought as that is a
SHOULD NOT rather than a MUST NOT, but this is why the schema
description notes that it is the _most_ compliant config,
and other configs may still be compliant.

Similary, the networknt Java implementation having `nullable`
on by default is not technically non-compliant, but it is more
compliant to not have it.

There may be other compliance configuration changes that need
documenting, but these were the ones that I could find with
a reasonable amount of effort.