Add last-updated tracking.
Since I just went through and looked at all of these, this adds
a date stamp for such things.  It is required throughout the data
file if it is used at all, but it can be left out entirely
(because I am not going to look through the obsolete files).

All four data files (validator/hyper + modern/obsolete) are valid
against the schema as of this commit.