Add built-on field
Some implementations, particularly but not exclusively CLI and
web ones, are implemented on top of other validator libraries.

This adds a structured field to account for that, and updates
all known cases where it is relevant.

It was necessary to also support an anchor-name field for
implementations as with programming languages, and in the
process disambiguate a few duplicate names.